+++ CHANGES.md
@@ -26,13 +26,13 @@
   - `some` - fallback on alternate non-nil values
 
   - Use data readers by default in `keypin.core/edn-file-io`
-- [Todo] Add support for lazy (late binding) references
+- Add support for lazy (late binding) references
   - New type `keypin.type.Ref`
   - Data readers `ref`, `ref!` - define key/path references
-  - [Todo] Augment `:parser` kwarg in `keypin.core/make-key`
+  - Augment `:parser` kwarg in `keypin.core/make-key` and lookup functions
     - Auto-resolve key/path references
     - Impacts `keypin.core/defkey`
-  - [Todo - Breaking] Lookup fns now invoke parser as `(fn [data key value])->value`
+    - [BREAKING] Lookup fns now invoke parser as `(fn [data key value])->value`
   - New fn `keypin.util/resolve-ref` for resolving references
 - [Todo] Deprecate `keypin.util/clojurize-subst` in favour of `keypin.util/resolve-ref`
 - [Todo] Reduce verbosity of config source lookup