+++ CHANGES.md
@@ -18,7 +18,7 @@
   - (Map) `^merge`   - Merge
 
 
-## [WIP] 0.8.3 / 2021-June-??
+## 0.8.3-beta1 / 2021-June-20
 
 - Add Keypin EDN data readers at `keypin.util/data-readers`
   - `env`, `env!` - to read environment variables